Ensure TTL is integer

+++ inc/pantheon-page-cache.php
@@ -422,7 +422,7 @@ class Pantheon_Cache {
 	 */
 	private function get_cache_control_header_value(): string {
 		if ( ! is_admin() && ! is_user_logged_in() ) {
-			$ttl = apply_filters( 'pantheon_cache_default_max_age', absint( $this->options['default_ttl'] ) );
+			$ttl = (int) apply_filters( 'pantheon_cache_default_max_age', absint( $this->options['default_ttl'] ) );
 			if ( $ttl < 60 && isset( $_ENV['PANTHEON_ENVIRONMENT'] ) && 'live' === $_ENV['PANTHEON_ENVIRONMENT'] ) {
 				$ttl = 60;
 			}
